requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
build==1.3.0
ccxt==4.1.74
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
numpy==1.24.4
packaging==25.0
pandas==2.1.4
propcache==0.3.2
pycares==4.10.0
pycparser==2.22
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
pytz==2023.3
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
TA-Lib==0.6.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.18
Werkzeug==3.1.3
yarl==1.20.1
